Move db connection info into dotenv
