+++ .env.example
@@ -1,0 +1,6 @@
+DB_DIALECT=
+DB_HOST=
+DB_PORT=
+DB_USERNAME=
+DB_PASSWORD=
+DB_NAME
