+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules
 keys
 .DS_Store
+
+.env
