+++ config/config.json
@@ -5,19 +5,5 @@
     "database": "fc5y_dev",
     "host": "127.0.0.1",
     "dialect": "mysql"
-  },
-  "test": {
-    "username": "root",
-    "password": null,
-    "database": "fc5y_test",
-    "host": "127.0.0.1",
-    "dialect": "mysql"
-  },
-  "production": {
-    "username": "root",
-    "password": null,
-    "database": "fc5y_prod",
-    "host": "127.0.0.1",
-    "dialect": "mysql"
   }
 }
