+++ models/user.js
@@ -1,6 +1,7 @@
 "use strict";
-const { Model } = require("sequelize");
-module.exports = (sequelize, DataTypes) => {
+const { Model, DataTypes } = require("sequelize");
+
+module.exports = (sequelize) => {
   class User extends Model {
     /**
      * Helper method for defining associations.
