+++ package.json
@@ -21,6 +21,7 @@
   "homepage": "https://github.com/fc5y/user-backend#readme",
   "dependencies": {
     "body-parser": "^1.19.0",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "mysql2": "^2.2.5",
     "sequelize": "^6.3.5"
